requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.4.1
billiard==3.3.0.22
celery==3.1.20
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.2
django-filter==0.12.0
django-haystack==2.4.1
django-redis-cache==1.6.5
djangorestframework==3.3.2
docopt==0.6.2
elasticsearch==2.2.0
eventlet==0.18.3
greenlet==0.4.9
hiredis==0.2.0
kombu==3.0.33
Markdown==2.6.5
psycopg2==2.6.1
pytz==2015.7
PyYAML==3.11
redis==2.10.5
requests==2.9.1
simplejson==3.8.1
static3==0.6.1
urllib3==1.14
